Update installation.md
Adressed several installation and running issues for running the server in a virtual environment on Apple Silicon Macs.

the Claude desktop config must be adjusted to point to the full path of uv for "command", assuming the target has installed uv only in the virtual environment. (I have uploaded a template that just works as long as anyone follows the guide.)

also, several dependency handling issues are still there. the persistent one is the enforcement of torch 2.5.1 while the uv.lock doesn't handle the installation of torch on Apple Silicon Macs correctly. That would be nice if you could resolve that. for the time being, I removed the dependency from uv.lock file and manually installed Pytorch dependencies. (The latest torch 2.6.0 seems work pretty well).